requirements.txt
sentence-transformers
faiss-cpu
flask
flask-cors
torch
numpy
transformers
gunicorn

